JAVAFIXTURE-16: Preliminary refactoring: merged AbstractSpecimenBuilder with SpecimenBuilder and simplified JavaFixture-class
